% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbcPMM.R
\name{AbcPMMLogLikCpp}
\alias{AbcPMMLogLikCpp}
\title{Calculate the AbcPMM log-likelihood for a given tree, data and model parameters using the Rcpp module}
\usage{
AbcPMMLogLikCpp(x, tree, x0, sigma2, sigmae2,
  cppObject = NewAbcPMMCppObject(x, tree),
  mode = getOption("SPLITT.postorder.mode", 0))
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}

\item{x0}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding white noise;}
\item{sigma2}{unit-time variance increment of the heritable component;}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{sigma2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding white noise;}
\item{sigma2}{unit-time variance increment of the heritable component;}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{sigmae2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding white noise;}
\item{sigma2}{unit-time variance increment of the heritable component;}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{cppObject}{a previously created object returned by \code{\link{NewAbcPMMCppObject}}}

\item{mode}{an integer denoting the mode for traversing the tree, i.e. serial vs parallel.}
}
\value{
the log-likelihood value.
}
\description{
Calculate the AbcPMM log-likelihood for a given tree, data and model parameters using the Rcpp module
}
