% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbcPMM.R
\name{NewAbcPMMCppObject}
\alias{NewAbcPMMCppObject}
\title{Create an instance of the Rcpp module for a given tree and trait data}
\usage{
NewAbcPMMCppObject(x, tree)
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}
}
\value{
an object to be passed as argument of the \link{AbcPMMLogLikCpp} function.
}
\description{
Create an instance of the Rcpp module for a given tree and trait data
}
\seealso{
\link{AbcPMMLogLikCpp}
}
